feat(core): expose `SafePathBuf`
