+++ .changes/safepathbuf-refactor.md
@@ -1,0 +1,5 @@
+---
+"tauri": patch
+---
+
+Expose `SafePathBuf` type in `tauri::path`.
