+++ core/tauri/src/api/file.rs
@@ -8,58 +8,11 @@
 mod extract;
 mod file_move;
 
-use std::{
-  fs,
-  path::{Display, Path},
-};
+use std::{fs, path::Path};
 
 #[cfg(feature = "fs-extract-api")]
 pub use extract::*;
 pub use file_move::*;
-
-use serde::{de::Error as DeError, Deserialize, Deserializer};
-
-#[derive(Clone, Debug)]
-pub(crate) struct SafePathBuf(std::path::PathBuf);
-
-impl SafePathBuf {
-  pub fn new(path: std::path::PathBuf) -> Result<Self, &'static str> {
-    if path
-      .components()
-      .any(|x| matches!(x, std::path::Component::ParentDir))
-    {
-      Err("cannot traverse directory, rewrite the path without the use of `../`")
-    } else {
-      Ok(Self(path))
-    }
-  }
-
-  #[allow(dead_code)]
-  pub unsafe fn new_unchecked(path: std::path::PathBuf) -> Self {
-    Self(path)
-  }
-
-  #[allow(dead_code)]
-  pub fn display(&self) -> Display<'_> {
-    self.0.display()
-  }
-}
-
-impl AsRef<Path> for SafePathBuf {
-  fn as_ref(&self) -> &Path {
-    self.0.as_ref()
-  }
-}
-
-impl<'de> Deserialize<'de> for SafePathBuf {
-  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-  where
-    D: Deserializer<'de>,
-  {
-    let path = std::path::PathBuf::deserialize(deserializer)?;
-    SafePathBuf::new(path).map_err(DeError::custom)
-  }
-}
 
 /// Reads the entire contents of a file into a string.
 pub fn read_string<P: AsRef<Path>>(file: P) -> crate::api::Result<String> {
@@ -76,19 +29,6 @@ mod test {
   use super::*;
   #[cfg(not(windows))]
   use crate::api::Error;
-  use quickcheck::{Arbitrary, Gen};
-
-  use std::path::PathBuf;
-
-  impl Arbitrary for super::SafePathBuf {
-    fn arbitrary(g: &mut Gen) -> Self {
-      Self(PathBuf::arbitrary(g))
-    }
-
-    fn shrink(&self) -> Box<dyn Iterator<Item = Self>> {
-      Box::new(self.0.shrink().map(SafePathBuf))
-    }
-  }
 
   #[test]
   fn check_read_string() {
